+++ src/polymer/behavior.ts
@@ -50,6 +50,7 @@ export class ScannedBehavior extends ScannedPolymerElement {
         getBehaviors(this.behaviorAssignments, document);
     const behavior = Object.assign(new Behavior(), this);
     behavior.warnings = behavior.warnings.concat(behaviorsAndWarnings.warnings);
+    this.applyJsdocDemoTags(document.url);
     return behavior;
   }
 }